rename molotov
